storage-operator/volume: fix newPersistentVolume
Replace hardcoded values for size and path by the ones from the volume.

On the other hand we hardcode the volume mode to `Filesystem` because
it's the only mode we support for now (note the use of a local variable
to workaround the fact that you can't take the address of a constant).

Refs: #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>