storage-operator/volume: implement volume creation
To "deploy" a volume, we need to prepare its storage (using Salt) and
create a backing PersistentVolume.

If we have no value for `Job`, that means nothing has started, and thus
we start the volume preparation using an asynchronous Salt call (which
gives us a job ID) and then reschedule the request to monitor the
evolution of the job.

If we do have a job ID, then something is in progress and we monitor it
until it's over.
If it has ended with an error, we move the volume into a failed state.
Otherwise we proceed with the PersistentVolume creation (which may
require an extra Salt call, synchronous this time, to get the disk
size), taking care of putting a finalizer on both ourself and the
PersistentVolume (so that our lifetimes are tied) and setting ourself as
the owner of the PersistentVolume.

Once we have successfuly created the PersistentVolume, we can move into
the `Available` state and reschedule the request (the next iteration
will check the health of the PersistentVolume we just created).

Here's is a little flow chart to help:

                    +------------------+             +----------+
              +---->|SpawnPrepareVolume|------------>|SetPending|
              |     +------------------+             +----------+
              | NO                                        |
              |                                           v
 -----     +----+ DONE  +--------+   +------------+   +-------+    ------
(START)--->|Job?|------>|CreatePV|-->|SetAvailable|-->|Requeue|-->( STOP )
 -----     +----+       +--------+   +------------+   +-------+    ------
             | YES                                         ^
             v                                             |
        +-----------+ Job Failed    +---------+            |
        |           |-------------->|SetFailed|----------->+
        |           |               +---------+            |
        |           |                                      |
        |           | Unknown Job   +--------+             |
        |PollSaltJob|-------------->|UnsetJob|------------>+
        |           |               +--------+             |
        |           |                                      |
        |           | Job Succeed   +--------+             |
        |           |-------------->|Job=DONE|------------>+
        +-----------+               +--------+             |
             | Job in progress                             |
             |                                             |
             +---------------------------------------------+

Refs: #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>