storage-operator/volume: check the PV owner
In order to detect a case where a user created a PV `foo` and then a
volume `foo` we shouldn't consider that the PV belongs to the Volume,
even though their names match because we haven't created the PV (and
thus Salt wasn't called to prepare the storage, …).
To detect these cases, we check that the owner of the PV is our volume,
if not we raise an error.

Refs: #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>