storage-operator/salt: handle disk.dump failure
It looks like executing `disk.dump` on a non-existing path doesn't
result in a direct failure but in a response like

    {
      "return": [
        {
          "bootstrap": false
        }
      ]
    }

Which makes Go `panic` when we try to cast `bootstrap` into a dict
(because here it's a bool).
Let's handle that case smoothly.

Note that this code is now failing because after taking into account
reviewers' comments I now use the correct path for the volume
(`/dev/disk/by-id/`) for both Volume type AND I always call Salt to get
the size: but because Salt isn't actually doing anything the code is now
failing (because the path doesn't exist).

That's the expected behavior so it's fine, but it makes things harder to
test until we have Salt working…

Refs: #1461, #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>