storage-operator/volume: cleanup logging context
In some place we were using `volume.Name` as `Request.Name` (which is
not 100% wrong but still can be confusing), at some other places we were
passing a `reconcile.Request` just for the sake of logging…), etc.
Let's only use `Volume.Name` everywhere it's possible: makes the code
more consistent and simpler.

Refs: #1462
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>