salt/volume: rework volume Salt module
We reworked the internals of the volumes Salt module in order to:
- make the SLS simpler (no more dispatch on the volume type with
  if/else)
- make it easier to support more volume types in the future ("just"
  implement a new class)

To do so, the module now expose a type-agnostic interface of free
functions wich provides the basic required actions (`is_formatted`,
`create`, …).

Then we have an abstract base class that can be inherited for each
Volume type to implement those actions as needed.

Refs: #1474
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>