storage-operator/salt: fix job polling
Until now, we used to rely on the value of `success` to know if a job
was completed successfully or not.
Unfortunately, that's not good enough: you can get `success: true` while
nothing was actually run…
See https://github.com/saltstack/salt/issues/4002

This is a problem because if we think that the `PrepareVolume` step was
a success, we will then try to get the size of the device with
`disk.dump`, but if `PrepareVolume` didn't run then the device doesn't
exists and `disk.dump` will fail => we wrongly move the Volume into a
failed state.

And we easily encounter this issue by creating several volume at
once/within a short time interval.

Try to spawn two state executions at the same time:

    salt-call state.sls metalk8s.volumes saltenv=metalk8s-2.4.0-dev pillar='{"volume": "foo"}'&
    salt-call state.sls metalk8s.volumes saltenv=metalk8s-2.4.0-dev pillar='{"volume": "bar"}'

You'll get:
    local:
        Data failed to compile:
    ----------
        The function "state.sls" is running as PID 32544 and was started at 2019, Aug 14 21:45:18.338129 with jid 20190814214518338129

And when it's done by the operator through the API, here is what we get
for the second job:

    "bootstrap": {
      "return": [
        "The function \"state.sls\" is running as PID 22874 and was started at 2019, Aug 14 20:47:46.622817 with jid 20190814204746622817"
      ],
      "retcode": 1,
      "success": true,
      "out": "highstate"
    }

`success` is true, but `retcode` is non-zero…

So, to be 100% sure that the spawned job succeeded we need to check the
return code.

Why not ignoring `success` and only rely on `retcode`?
Because we want to tell the difference between:
- the job ran but failed (success == false): move the Volume into Failed
  state
- the job failed to run (success == true but retcode != 0), probably
  because one was already running: we can retry later

Note that it seems possible to run states in parallel[1], but our state
may install packages and that seems to be a dangerous things to do in
parallel, so let's avoid…
(But maybe we should move the deps install out of this state, that would
probably make sense…)

[1]: https://docs.saltstack.com/en/latest/ref/states/parallel.html

Refs: #1474
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>