salt/volume: use UUID for `is_formatted`
If we want to be really sure that a device have been formatted by use we
shouldn't check the FS type but the UUID.

In fact, relying on the FS type for `is_formatted` can be wrong.

- a Volume V is created with the StorageClass SC which has `fsType` set
  to `xfs`
- we format the device with XFS
- later, the StorageClass SC is deleted and someone reuse the name to
  create a new StorageClass SC which has `fsType` set to `ext4`
- the server reboot, we reprovision all the volume declared on the node,
  when the pillar is updated it stored the volume V and embed the
  content of the (new) StorageClass SC
- when we check if the device FS type (xfs) match the expected one (from
  the StorageClass, now `ext4`) we get a mismatch and think the device
  is not formatted
- we then call `format`, here the safeguard detect that the device is
  already formated and we bail out and avoid data loss, but we have put
  the Volume V in a Failed state when everything was fine: that's not
  good…

So yeah, we really should really on the UUID instead of the FS type.

Note that this only hold for `is_formatted`, the safeguard in `format`
still need to check if we already have something on the device before
formatting to avoid data loss.

Refs: #1474
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>