salt/volume: use libblkid directly
This commit adds a new Python module to call directly into the blkid
library (using ctypes) instead of relying on the `blkid` command.

Why do we do that?
==================

Because we want to be as sure as possible that when `blkid` returns no
informationg is means "there is no info" and not "maybe there is some
info but I couldn't get them" (because if we mix two cases, we may
actually format something because we thought "ho there is nothing,
davai" whereas the reality is "there was data but I haven't seen them
because reasons…").

Relying on `return code 2 == no data` seems kinda safe, but it's not
100% clear from the man page. Now we could check the source code  (and I
did) but that would only be valid for the current version we look at.

So let's directly attack the library, making sure that any API call will
result in an exception and thus if we end up with "no data" then that
truly means "everything went well, but there was nothing there".

Refs: #1474
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>