doc/dev: move local cluster instructions
We now have several component that can be run locally (a whole cluster,
the UI, the storage operator, …) and thus it makes sense to have a
dedicated section where to put group them.

Since we now have a dedicated section, this commit moves the
instructions on how to run a local cluster into this new section.

Refs: #1536
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>