test: relax Gherkin language validation globally
Because having implicit dependencies between test scenarios is pretty
bad (trust me, been there done that…) we want to make test scenarios
self-contained (as much as possble).

And to do so, we sometimes need to have more than one sequence of
`When/Then` in a single scenario, but `pytest.bdd` (unlike `behave`…)
forbids that by default.

So let's relax the validation to allow this kind of pattern.

Refs: #1417
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>