salt/volume: add volume cleanup to our Salt state
We add a new formula to clean up a MetalK8s volume.

For rawBlockDevice, we're doing nothing (no `wipefs`) so that if the
user deleted the volume by mistake, there is no data loss (and if they
really want to reuse the device they can format it explicitly by
themselve).

For sparseLoopDevice, we do remove the loop device and delete the
backing sparse file, otherwise they would keep piling up in the
`/var/lib/metalk8s/storage/sparse/` directory.
Moreover, it's only used for CI/demo purpose, so we don't need to worry
that much about data loss.

Refs: #1475