storage-operator: use leader-with-lease mode
By default, the operator use a `leader-for-life` election mode which has
the nice property of avoiding split brain.
Unfortunately, when the operator's pod is killed by Kubernetes (because
resources are becoming scarce), the dead pod doesn't step down as a
leader (because evicted Pod are not deleted, thus they keep a reference
on the leader lock that then can't be garbage-collected) and the new one
spawned by Kubernetes is stuck with messages like:

    […]
    Trying to become the leader.
    Found existing lock (LockOwner: storage-operator-56f98777b9-mr9lp)
    Not the leader. Waiting.
    Not the leader. Waiting.
    […]

This is a known bug: see
https://github.com/operator-framework/operator-sdk/issues/1305

As a workaround, you can manually deleted the evicted pod (but this
involve manual intervention and is not acceptable).

Now, there is a second mode of leader election: "leader-with-lease".
This one requires a periodic renewal of the leader lease and gives up
leadership when the lease is not renewed: that way, an evicted pod will
eventually relinquish it's leadership and the freshly spawned one will
be able to take over.
Note that we can now end up with split brain, which should be fine in
our case because Salt is a serialization point (you can't run two Salt
states to prepare or finalize a Volume), thus we won't run two states
concurrently, and running the same state twice in a row on the same
volume is fine thanks to idempotency.

So, until the bug in the Operator SDK is fixed, let's use
"leader-with-lease" election mode.

[1]: https://docs.openshift.com/container-platform/4.1/applications/operator_sdk/osdk-leader-election.html

Closes: #1580
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>