test/volume: add a fixture to cleanup volumes
To avoid unwanted side-effect and unexpected interactions between test
scenarios, we don't want volumes created into one scenario to be still
alive when the next one is executed.

To see why it's an issue, consider the latest test we introduced, where
we create PVC and expect it to claim the volume we just created. If old
volumes still lives around, the PVC may claim one those instead of the
one we just created => the test will now randomly fails and that's not
good.

Thus, we introduce a fixture which will cleanup any dangling volumes at
the end of each scenario.

Refs: #1417
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>