test/volume: add pod & pvc cleanup
Some tests may leave dangling pods behind, and if those pods are using a
volume, they will prevent the cleanup of the volume.

Ditto for the PersistentVolumeClaim.

Thus, we first cleanup the pods, then the claims and finally we cleanup
the volumes.

Note that we renamed the fixture as it's no longer only about Volumes.

Refs: #1417
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>