build/package: add an entrypoint for the Debian builder
This commit add a Shell script that will be used by the Debian builder
container to build DEB packages.
It supports:
- the creation of a DEB package from an RPM one, using alien
- the creation of a DEB package from the sources, using debuild

Note that we specify the version in the changelog instead of inside the
control file because counterintuitively `dpkg-buildpackage` infers the
package version number from this file, and not from a "Version" field in
control.

Refs: #1472