storage-operator/volume: ignore failed JID
When we're trying to delete a job in a failed state, it may already have
a Job ID (if the Volume failed during its creation) that we kept around
for troubleshooting purpose (one may want to manually query Salt API
with the Job ID to get more info about why it failed).

Now, we must ignore this Job ID otherwise we will wrongly think that the
UnprepareVolume is already in progress…
So let's ignore Job ID from Failed Volume, once we start to unprepare
the volume we will get our own Job ID and move into the Terminating
state (so that the Job ID won't be ignored in the next iterations).

Refs: #1593
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>