salt/volume: fix unprepared state
If a volume is not found in the pillar, we shouldn't fail. Instead we
should consider that we have nothing to do.

Indeed, consider what happens if you try to create a Volume which is not
in the pillar (because it uses a non-existing storage class for
instance): this Volume will fail with a CreationError and when you'll
try to delete this Volume you've nothing to unprepare since it was never
created to begin with.
In such case, you don't want to have unprepare failing.

Refs: #1593
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>