salt/volume: don't filter out volume from pillar
We used to filter out the Volume without an associated StorageClass that
exists.
The problem is that if you delete a StorageClass after creating a volume
it should be perfectly OK (because we only use it at creation time), but
that's not true: even though the `unprepared` state doesn't use the
StorageClass, it does need the Volume. But since the Volume reference a
non-existing StorageClass, it's not present in the pillar and thus
cannot be cleaned up by the `unprepared` state.

To avoid this kind of situation we no longer filter out Volume
referencing a non-existing storage class. Instead, we store the Volume
in the pillar, but without the StorageClass information.
Then, the `prepared` state will check if we do have StorageClass
information, and if not it will raise an error.
Since `unprepared` doesn't care about the StorageClass, nothing will
happen there and we will happily clean up the Volume (because this time,
we will have an entry in the pillar).

Refs: #1593
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>