test/volume: check backing storage
In order to be sure that the Salt states are working correctly, we had
check on the backing storage.

Note that to check that the backing storage is deleted, we need the
Volume UUID but we can't get it as the Volume as already been deleted at
that point. To solve this problem, we add a context where the functions
that fetch the volume can store the volume info which then can be reused
by later steps.

`_quantity_to_bytes` is copied from the Salt module (really not fond of
code duplication, but dunno how to share code between test suite and
Salt tree…).

Refs: #1593
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>