build/packaging: add task to build the DEB packages
In order to build the required DEB packages, we create a new task
generator, `_build_deb_packages`, that will create a sub-task for each
DEB package to build.

This task relies a a list of pre-defined packages, DEB_TO_BUILD.

Given that some DEB packages depends on the RPM one, we extract into
dedicated constants what is shared between DEB and RPM build tasks:
namely the calico package and the scality repository.

Refs: #1472