build/packaging: Remove auto generated file
During the Debian package generation tests, a file named `file` has
been created. This file should not appear in the list of the source
files and will be not because the volume which will be mounted in the
container to build the Debian package is on `read-only`, so this file
can be remove.

Refs: #1472