salt: provide a cryptconfig state for apiserver
This state is responsible for generating the configuration file that
`kube-apiserver` takes as argument of its launch command.

If the `encryption` key is unavailable from the pillar, the configuration
file is generated by the state. This is guaranteed to be on the bootstrap
node due to the fact that it is always the first node to be brought to
highstate.