package/debian: compute the list of packages to download
From each package we want to explicitly install, we walk the dependency
tree and fetch all the required dependencies (ignoring duplicates).

As for version selection, if one is specified (can be the case for some
packages we explicitly install) we use that one, otherwise we select the
most recent available.

Refs: #1668
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>