monitoring: use persistent storage for Prometheus
These changes configure the Prometheus stateful services (i.e.,
Prometheus and AlertManager) to use a persistent volume for storage
instead of ephemeral `emptyDir` volumes.

As a result, these services don't 'work' out of the box after initial
deployment. We add some example `Volume` templates using the
`sparseLoopDevice` storage type to create proper volumes on a node after
deployment, which should allow the services to start.

The `metalk8s-prometheus` *StorageClass* that used to be provisioned as
part of the `volumes` add-on is now deployed as part of the
`prometheus-operator` add-on. The rationale is as follows: the `volumes`
add-on (i.e., the operator we deploy) doesn't require any particular
*StorageClass* to be present, and as such should not deploy any
*StorageClass*es with it. Instead, an application which consumes
*PersistentVolume*s of a particular *StorageClass* must deploy these
*StorageClass*es with it, in casu the `prometheus-operator` add-on.

Furthermore, the Vagrant deployment and CI are extended to create the
required `Volume`s such that the Prometheus services *are* available
after bootstrapping the system.

Note: during an earlier implementation of this feature (#1523), we ran
into a dependency issue between Prometheus deployment and the MetalK8s
UI deployment, which requires the first to be available. Instead, given
this PR, we use a hard-coded `NodePort` port number to access
Prometheus, not requiring the service to be available during bootstrap,
and point the UI to the workload-plane IP of the bootstrap node. This
ought to be improved once we use an `Ingress` to access the UI.

Fixes: #1515
See: https://github.com/scality/metalk8s/issues/1515
See: https://github.com/scality/metalk8s/pull/1523