versions: introduce common packages and extras/overrides
The support of an additional OS requires a 'compatibility' layer for
package names and versions.

We define a common core of packages whose names do not diverge across OS
families and whose versions are either latest or fixed across OS families.

We introduce overrides in package versions, such that a package for a specific
version may override the package referenced in the common core.

We unpin most packages in the RPM listing, where we only really need to pin
K8s core components as well as some of the 'unstable' Python APIs we rely
on.

Cherry-picked from eb2afe719deafb5338135c30a008c3f91f317ce6

Issue: #1530