test/lint: fix shell scripts listing
We were building a list of shell scripts to lint using glob pattern.
The problem is that `**` is not recursive with Bash (it is with zsh, but
we are using bash for tox).
Also, even if it was recursive that would be no good because it would
select files we don't want to check (like stuff under `node_modules`,
e.g. `ui/node_modules/coa/lib/completion.sh`).

Now, we are relying on the output of `git ls-files` which handle
`.gitignore` and list everything without needing a recursive glob
pattern.

Refs: #1637
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>