test/lint: fix YAML listing
Same problem as shell scripts linting: the glob pattern is not
recursive.
And if it was, it would be broken because the Salt tree contains `.yaml`
that aren't plain YAML but Jinja-templated one (which `yamllint`)
doesn't like.

In the end, there is only `salt/metalk8s/defaults.yaml` as pure YAML
under `salt/` so let's lint this one explicitely.

Refs: #1637
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>