build/targets: add support for source prefix in FileTree
Until now, the source path of the files to copy was used to compute its
destination path.

e.g.: with a destination set to 'foo' and a file 'bar/qux', then the
file was copied to 'foo/bar/qux'

This convention was convenient, but also limitating. To circumvent it,
we introduce a new optional parameter, `source_prefix`, which is
prepended to each source filepath BUT is ignored to compute the
destination path.

With this parameter we now support copying `bar/qux` to `foo/qux` with:
- destination: `foo`
- file: `qux`
- source prefix: `bar`

Refs: #1439
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>