build/packaging: Change the Debian packages location
When building repositories, the packages are copied into the repo
hierarchy.

Until now, we were downloading package on the ISO directly (under
`_build/root`) but since the repo copy them, we shouldn't download them
here (they will be duplicated).

This commit updates the code to download the packages in the "staging"
area (under `_build/packages`).

Refs: #1746