build/doc: allow to reuse pre-build doc
In the CI, the documentation is built by another container/step and
during the build step we want to reuse this artifact instead of
re-building it ourself.

This commit introduce a new config knob to check if we are running in
the CI or not.

This is is a kinda internal settings, we don't document it outside of
the code.

Refs: #1695
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>