storage-operator/volume: add `Conditions` fields
This commit introduces a new field in `VolumeStatus`: `Conditions`. This
field contains a list of `VolumeCondition` and supersedes the fields
`Phase`, `ErrorCode` and `ErrorMessage`.

Using a list of conditions allows for more flexibility (we can add new
conditions without impacting the CRD) and is the canonical way to
represent status in the Kubernetes world.

Refs: #1507
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>