storage-operator: adapt code to use conditions
This is the first step toward using the conditions.
We won't harvest the benefits of using conditions yet, this commit is
mainly to introduce the schema changes while keeping the old logic of a
"global" status.

For now we are using a single condition, Ready, which can be used to
compute a "global" status and allows us to keep most of the code
unchanged.

The mapping is as follow:

    | `status.phase` | `Conditions[Ready]`                   |
    | -------------- | ------------------------------------- |
    | `Unknown`      | missing                               |
    | `Failed`       | `False`                               |
    | `Pending`      | `Unknown` + `Reason` ==  `Pending`    |
    | `Available`    | `True`                                |
    | `Terminating`  | `Unknown` + `Reason` == `Terminating` |

Refs: #1507
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>