ui/volume: adapt code to use conditions
Still striving to minimize the changes, we only fix the creation of the
Volume object in the UI (in `NodeInformation.js`) by using a function to
compute a global status.
This allows us to keep most of the code unchanged (no need to rework the
table or any visible component), which reduces the risk of regression.

Closes: #1507
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>