build/image: rely on Docker API for uptodate
Until now, we were relying on a `touch`ed file on disk to know if we had
built the Docker image.
This is perfectible™ and doesn't handle the case where one delete
manually the image through Docker.

This commit adds a custom `uptodate` predicate that checks if the image
is present in the local Docker registry, which is more reliable.

Closes: #839
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>