build/docker: rework the default run config
We can no longer relies on `get_entrypoint_mount` because we will have
entrypoint unrelated to packaging.

So, instead of using a set of constant for the default configuration we
introduce a new helper that generate a default configuration from the
entrypoint path (for packaging entrypoints we had new constants because
they are referenced in several files).

Refs: #1830
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>