build/builder: add a builder for the Go code
Like for documentation, avoid requiring people to install Go (and the
deps for the Operator SDK) if they just want to build the ISO (and not
actually hack on the code).
Will also ensure that everyone use a single version of Go to build,
which avoids weird discrepencies (see Go 1.12 vs Go 1.13 behavior w.r.t.
the fix for the thrift apache repository…)

Refs: #1831
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>