storage-operator/build: add an entrypoint script
If we want to build the docs inside a container, we must chown back
the generated files, otherwise they will be owned by root.

Note that since we are using an Alpine image, we are using `sh` (`bash`
is not installed by default), but given the simplicity or our script
that's not an issue.

Refs: #1831
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>