build: generate a build.log
We update our custom reporter to generate JSON build log at the end of
the build.

To do so, we inherit from `JsonReporter` instead of `ConsoleReporter`
and we simply patch `outsream` after the build is complete in order to
write the JSON log into a file instead of `stdout`.

In the end, our CustomReporter is a mix between `ConsoleReporter` and
`JsonReporter`, with a twist (write JSON dump into a file).

Refs: #1849
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>