ci: upload the build.log to artifacts
We now upload the build.log to artifacts for future processing.
Since what we upload is not only related to the ISO (heck, even the step
is called "upload artifacts", not "upload iso") we rename the directory
`artifacts`.

The line being too long, let's use several cp command, that won't kill
us.

Closes: #1849
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>