ci: workaround a CI bug
The CI only watch the content of the `context` directory to know when to
rebuild the images, so if you have a Dockerfile out of the context
directory and that you only modify the Dockerfile (that's what we do in
this branch for the `lint` and `unit-test` images), then you images
won't be rebuilt.

A fix will be deployed soon, meanwhile we work around this behavior by
adding a dummy comment in `go.mod` (which is in the `context`).

See EVE-1098

Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>