doc/deps: require sphinxcontrib-plantuml > 0.16.1
The doc build sometimes fails with:

    Exception occurred:
      File "[…]/sphinxcontrib/plantuml.py", line 421, in latex_visit_plantuml
        self.builder.warn(str(err))
    AttributeError: 'LaTeXBuilder' object has no attribute 'warn'

But the real error (in `err`) is hidden by the AttributeError…

This was fixed in
https://github.com/sphinx-contrib/plantuml/commit/729138f9e7a3c7a74602fcd9eb55f9cec846be1b,
which was released in 0.16.1

So let's update our version of sphinxcontrib-plantuml (and bump the
minimal required version).

Hopefully, next time the doc build fails we will get the real error and
find a way to fix it!

Refs: #1872
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>