Salt/repo: Deploy the Ubuntu repositories
Permit to deploy the internal repositories sources on the Ubuntu machine
and to give the highest priority to these repositories to be sure that
we use these ones. The option to disable official repo doesn't exist
on Ubuntu, so we use the repo priorities to have the same required effect.

Ref: #1728