ci/build: allocate more RAM for the build step
At the time we defined those memory request, we had a single container
running at once: the RPM builder for CentOS packaging.

In 2.4 we've added:
- a container for Ubuntu packaging
- a container to build the storage-operator
- a container to build the documentation

And those, being independent, can run in parallel and when it happens we
may need more than what was requested here.

Let's bump the requested memory  by 1G, this should reduce the doc build
failures.

Closes: #1872
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>