salt/solutions: Revamp configuration file
A new field, currently named `active` (may be a poor choice, we'll see
with first user interactions, `activeVersions` could be more explicit),
is introduced to set which version of a Solution should be used for the
cluster-wide components of this Solution (i.e. the Admin UI and CRDs for
now).
Validation of the file format is added, using the usual `kind` and
`apiVersion` fields. Later improvements could use some OpenAPI schema.

Fixes: #1582