salt/solutions: Revamp current state in pillar
Previous approached relied on a ConfigMap, which poses numerous issues:
- it allows one to lie to the formula, potentially causing further
problems down the line
- the CM cannot always be up-to-date with the actual system state (due
to errors when updating it, or manual operations outside of Salt)

We change this approach by including basic heuristics in the
`metalk8s_solutions` module, which can be refined in the future.

Listing of mounted Solutions is made based on `mount.available` and a
naive filtering on the mountpoints (necessarily under /srv/scality).

Listing of active Solution versions is made by looking for UI Service
objects in the `metalk8s-solutions` namespace. This approach relies on
labels that we don't set ourselves, which is brittle, and does not
account for Solutions that may want to ship without an Admin UI.

Future iterations could improve those methods, but we are convinced it
is still safer than the ConfigMap method used before.

Issue: #1852