salt/solutions: Revamp orchestrate for K8s objects
Once a Solution is "available" (i.e. mounted and its images exposed by
the cluster registry), its cluster-wide components can be deployed.

We rely on the `active` mapping of active versions in the config file to
know which components to deploy or remove.

Note that one can set a desired version to 'latest' in this mapping,
which the orchestration will handle for using the highest version
available for a Solution (in SemVer terms).

Issue: #1852