scripts: Monkey-patch old versions of argparse
Old (<2.7.9) versions of `argparse` have a different behaviour when it
comes to defaults set by subparsers. We need the most recent behaviour
for injecting the selected `Command` class.

Alternatives to this monkey-patch were deemed too complex in comparison
(e.g. enforcing Python 3, with the implications in terms of packaging
for CentOS/RHEL 7).