build/repository: don't use `:` to emulate nested tasks
The semicolon has a special semantic in doit: it is used to separate the
task basename from the task name.

I wanted to have kind of task tree where you could:
- build all the repositories at once: `_build_rpm_repositories:*`
- build a single repository: `_build_rpm_repositories:scality:*`
- run a single step for every repository:
  `__build_rpm_repositories:*:build_repodata`

That's indeed a nice feature, but using `:` as delimiter was wrong
because it messed up doit behavior in a subtle way.

If you look linting (./doit.sh info lint), you can see that:
- you have a base task `lint`
- you have children tasks for each linting "module"
- the base task depends on the children tasks

But that doesn't work for tasks that uses this "nested" pattern: you
don't have a base task `_build_deb_repositories`…
This make the use of more advanced feature (such as delayed task
creation) more troublesome than it should be.

So let's get rid of this naming scheme. Now:
- we have a basename (e.g. `_build_deb_repositories`)
- then a semicolon
- then the task name, prefixed by the repository name.

No regression feature-wise, we can still:
- build all the repositories at once: `_build_rpm_repositories`
- build a single repository: `_build_rpm_repositories:scality_*`
- run a single step for every repository:
  `__build_rpm_repositories:*_build_repodata`

Refs: #1911
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>