build/packaging: use delayed task for build_deb_repositories
It seems that Ubuntu loves to move its packages around, and sometimes we
may end up with empty repository. Those shouldn't be built, and that's
the purpose of this commit: skip the build of empty repositories.

In order to know if a repository is empty or not, we first need to
resolve the dependencies and download the corresponding packages.
Now, `doit` works in two steps:
- first, it creates all the tasks
- then, it executes them

Here we want to create the `build_repository` after the execution of
`download_deb_packages` and that can be done using delayed task
creation[1].

[1]: https://pydoit.org/task_creation.html?#delayed-task-creation

Refs: #1911
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>