salt: set osmajorrelease in yum vars
This variable is available in all yum repo files and
based on the Salt grain with the same name.

We need this in order to handle CentOS & RedHat
the same way, because $osreleasever is different
in CentOS (7) and RedHat (7Server) and we're using
the same repositories, so the URL would not match.

Refs: #1552