vagrant/ubuntu: Add a repo update on Ubuntu VM
 - Context:
We need to install Python2.7 on the Ubuntu VM that we deploy with Vagrant.
Sometimes, some packages are moved to a Canonical repo to another one.
To avoid to have an error during the python2.7 installation, we want to
do an update.

 - Change(s):
Modification of the VagrantFile to do an update before to launch the
Python2.7 install.

Ref: #1931