scripts: Improve management of subprocess
The initial crude implementation relied on `check_output`. We now
implement a poor man's `subprocess.run`, with defaults suiting our use
case (namely, an equivalent of the `CalledProcessError` being raised for
loggers to catch and display, `stdout` and `stderr` being captured).
This will be applied to existing commands in a later commit.